Major changes in ldComputations.py. Details as follows:
Completely re-wrote the function PartitionDisEq. This wasn't an error
correction, but just a re-ordering of computations.
The earlier version of PartitionDisEq calculated the same final return
value, but ordered the computations differently. But the new way of
organizing (specifically, calculating the disequilibrium (the E score, see
the bottom of page 2, Dpca.pdf) as the average of individual row scores)
has the advantage that it can be used without
modification to calculate Marcy's individual haplotype scores (1. See
Section 2, Dpca.pdf. 2. Marcy also defines similar individual zygote
scores, but PartitionDisEq cannot be used to calculate the zygote
scores, except the ID disequilibrium.)

Also it can be seen that the new code is highly "vectorized". So hopefully
the numpy array computations will go fast.

As with the earlier version, I ran the script simple_level3_test.sh, which
(1) ran DQ.py to produce nomix1214.pop1.level3idd.out, and (2) ran
formatforR.py on the .out file to produce the
nomix1214.pop1.level3idd.rdata file.  

I did svn diff, and the differences in the scores between the previous and
the new version of nomix1214.pop1.level3idd.out occurs only at the 16th
(or so) decimal digit. I am putting this down to differences in the way the
numerical computations are ordered.